+++ R/draw.R
@@ -78,7 +78,7 @@ draw.gp <- function(object, type = 'rmse', log = FALSE, ...){
     wave_N <- length(object$design)
     if ( "type" %in% names(object$design) ) wave_N <- wave_N - 1
     if ( "exclusion" %in% names(object$design) ) wave_N <- wave_N - 1
-    cust <- ifelse ( is.function(object$design$type), T, F )
+    cust <- object$design$type=='customized'
     if ( "x_test" %in% names(object$design) & "y_test" %in% names(object$design) ) wave_N <- wave_N - 2
     for ( i in 1:wave_N ){
       Ni <- sum(object$design[[paste('wave',i,sep='')]]$enrichment)
@@ -200,7 +200,7 @@ draw.dgp <- function(object, type = 'rmse', log = FALSE, ...){
     wave_N <- length(object$design)
     if ( "type" %in% names(object$design) ) wave_N <- wave_N - 1
     if ( "exclusion" %in% names(object$design) ) wave_N <- wave_N - 1
-    cust <- ifelse ( is.function(object$design$type), T, F )
+    cust <- object$design$type=='customized'
     if ( "x_test" %in% names(object$design) & "y_test" %in% names(object$design) ) wave_N <- wave_N - 2
     for ( i in 1:wave_N ){
       Ni <- sum(object$design[[paste('wave',i,sep='')]]$enrichment)
@@ -342,7 +342,7 @@ draw.bundle <- function(object, emulator = 1, type = 'rmse', log = FALSE, ...){
     wave_N <- length(object$design)
     if ( "type" %in% names(object$design) ) wave_N <- wave_N - 1
     if ( "exclusion" %in% names(object$design) ) wave_N <- wave_N - 1
-    cust <- ifelse ( is.function(object$design$type), T, F )
+    cust <- object$design$type=='customized'
     if ( "x_test" %in% names(object$design) & "y_test" %in% names(object$design) ) wave_N <- wave_N - 2
     for ( i in 1:wave_N ){
       Ni <- sum(object$design[[paste('wave',i,sep='')]]$enrichment[,emulator])