+++ NEWS.md
@@ -16,6 +16,7 @@
 - `pack()` can now accept a list of (D)GP emulators as the input.
 - The `check_point` argument is removed from `design()` and replaced by `autosave`.
 - Automatic saving of emulators during the sequential design is added to `design()` through the new argument `autosave`.
+- When a customized evaluation function is provided to `design()` via `eval`, the design information in previous waves will be retained as long as the previous waves of the sequential design also use customized evaluation functions. If different customized evaluation functions are supplied to `design()` in different waves, the trace plot of RMSEs produced by `draw()` will show RMSEs from different evaluation functions in different waves.
 - Updates of documentations and vignettes.
 
 # dgpsi 2.3.0